% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure-clients.R
\name{api_graph_client}
\alias{api_graph_client}
\title{Microsoft Graph API Client}
\description{
An R6 class that extends \link{api_client} with Microsoft Graph-specific defaults.
This client is preconfigured with the Graph API host URL and scope.
}
\keyword{internal}
\section{Super class}{
\code{\link[azr:api_client]{azr::api_client}} -> \code{api_graph_client}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-api_graph_client-new}{\code{api_graph_client$new()}}
\item \href{#method-api_graph_client-clone}{\code{api_graph_client$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="azr" data-topic="api_client" data-id=".fetch"><a href='../../azr/html/api_client.html#method-api_client-.fetch'><code>azr::api_client$.fetch()</code></a></span></li>
<li><span class="pkg-link" data-pkg="azr" data-topic="api_client" data-id=".get_token"><a href='../../azr/html/api_client.html#method-api_client-.get_token'><code>azr::api_client$.get_token()</code></a></span></li>
<li><span class="pkg-link" data-pkg="azr" data-topic="api_client" data-id=".req_build"><a href='../../azr/html/api_client.html#method-api_client-.req_build'><code>azr::api_client$.req_build()</code></a></span></li>
<li><span class="pkg-link" data-pkg="azr" data-topic="api_client" data-id=".req_perform"><a href='../../azr/html/api_client.html#method-api_client-.req_perform'><code>azr::api_client$.req_perform()</code></a></span></li>
<li><span class="pkg-link" data-pkg="azr" data-topic="api_client" data-id=".resp_content"><a href='../../azr/html/api_client.html#method-api_client-.resp_content'><code>azr::api_client$.resp_content()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-api_graph_client-new"></a>}}
\if{latex}{\out{\hypertarget{method-api_graph_client-new}{}}}
\subsection{Method \code{new()}}{
Create a new Microsoft Graph API client instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{api_graph_client$new(
  provider = NULL,
  host_url = "https://graph.microsoft.com",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{provider}}{A credential provider for authentication. If NULL,
will be passed to \link{api_client}.}

\item{\code{host_url}}{The Microsoft Graph API base URL. Defaults to
"https://graph.microsoft.com".}

\item{\code{...}}{Additional arguments passed to \link{api_client} constructor.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{api_graph_client} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-api_graph_client-clone"></a>}}
\if{latex}{\out{\hypertarget{method-api_graph_client-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{api_graph_client$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
