% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure-clients.R
\name{api_graph_service}
\alias{api_graph_service}
\title{Microsoft Graph API Service}
\description{
An R6 class that extends \link{api_service} with Microsoft Graph-specific configuration.
This service is preconfigured with Graph API client and endpoints (v1.0 and beta).
}
\keyword{internal}
\section{Super class}{
\code{\link[azr:api_service]{azr::api_service}} -> \code{api_graph_service}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-api_graph_service-new}{\code{api_graph_service$new()}}
\item \href{#method-api_graph_service-clone}{\code{api_graph_service$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="azr" data-topic="api_service" data-id="print"><a href='../../azr/html/api_service.html#method-api_service-print'><code>azr::api_service$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-api_graph_service-new"></a>}}
\if{latex}{\out{\hypertarget{method-api_graph_service-new}{}}}
\subsection{Method \code{new()}}{
Create a new Microsoft Graph API service instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{api_graph_service$new(chain = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{chain}}{A \link{credential_chain} instance for authentication. If NULL,
a default credential chain will be created.}

\item{\code{...}}{Additional arguments passed to \link{api_graph_client} constructor.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{api_graph_service} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-api_graph_service-clone"></a>}}
\if{latex}{\out{\hypertarget{method-api_graph_service-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{api_graph_service$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
