% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az-dataset.R
\name{az_data_catalog}
\alias{az_data_catalog}
\title{Azure Storage dataset catalog}
\usage{
az_data_catalog(datasets = list())
}
\arguments{
\item{datasets}{A list of \link{az_dataset} objects.}
}
\value{
An \code{az_data_catalog} S7 object.
}
\description{
An S7 class holding an ordered collection of \link{az_dataset} objects with
unique \code{name}s.
}
