% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azr-package.R
\docType{package}
\name{azr-package}
\alias{azr}
\alias{azr-package}
\title{azr: Credential Chain for Seamless 'OAuth 2.0' Authentication to 'Azure Services'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Implements a credential chain for 'Azure OAuth 2.0' authentication based on the package 'httr2''s 'OAuth' framework. Sequentially attempts authentication methods until one succeeds. During development allows interactive browser-based flows ('Device Code' and 'Auth Code' flows) and non-interactive flow ('Client Secret') in batch mode.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pedrobtz.github.io/azr/}
  \item \url{https://github.com/pedrobtz/azr}
  \item Report bugs at \url{https://github.com/pedrobtz/azr/issues}
}

}
\author{
\strong{Maintainer}: Pedro Baltazar \email{pedrobtz@gmail.com}

Authors:
\itemize{
  \item Pedro Baltazar \email{pedrobtz@gmail.com}
}

}
\keyword{internal}
