% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azr-dataset.R
\name{azr_catalog_write}
\alias{azr_catalog_write}
\title{Write a dataset catalog to JSON}
\usage{
azr_catalog_write(catalog, json_file)
}
\arguments{
\item{catalog}{An \link{azr_catalog} object.}

\item{json_file}{Path to write the JSON file to.}
}
\value{
\code{json_file}, invisibly.
}
\description{
Writes an \link{azr_catalog} to a JSON file in the shape expected by
\code{\link[=azr_catalog_read]{azr_catalog_read()}}.
}
\seealso{
\code{\link[=azr_catalog_read]{azr_catalog_read()}}
}
