% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{azr_opt}
\alias{azr_opt}
\title{Read an azr package option}
\usage{
azr_opt(name)
}
\arguments{
\item{name}{A character string naming the option. Must be one of the keys in
\link{azr_options} (e.g. \code{"verbose"}).}
}
\value{
The option value, coerced to the type of the built-in default.
}
\description{
Retrieves the value of an azr package option by name, using the following
priority order:
\enumerate{
\item R option (e.g. \code{options(azr.verbose = TRUE)})
\item Environment variable (e.g. \code{AZR_VERBOSE=true})
\item Built-in default
}
}
\examples{
azr_opt("verbose")
}
