% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_vm.R
\name{azr_setup_vm}
\alias{azr_setup_vm}
\title{Set Up an Azure Virtual Machine for R}
\usage{
azr_setup_vm(
  name = make.names(Sys.info()["user"]),
  size = "Standard_D2s_v5",
  loc = "eastus",
  ssh_pub = NULL,
  dir = Sys.getenv("AZURE_DIR"),
  sub_id = Sys.getenv("AZURE_SUB_ID")
)
}
\arguments{
\item{name}{the name of the resource group. Defaults to your computer's username.}

\item{size}{Machine specs. Some recommended options:
* \code{Standard_B2s}: 2 cores, 4 GB RAM
* \code{Standard_D2s_v5}: 2 cores, 8 GB RAM
* \code{Standard_D4s_v3}: 4 cores, 6 GB RAM
See \url{https://learn.microsoft.com/en-us/azure/virtual-machines/sizes} for more information.}

\item{loc}{The physical location of the VM. Defaults to \code{eastus} in Virginia.}

\item{ssh_pub}{SSH public key, or path to a file containing it. By default
will search for a public key file in the usual location.}

\item{dir, sub_id}{Azure identifiers that should be set in your \code{.Renviron}. See
the package README for details.}
}
\value{
A reference to the VM, invisibly
}
\description{
Set Up an Azure Virtual Machine for R
}
