% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.R
\name{azr_ssh_cmd}
\alias{azr_ssh_cmd}
\title{Command to SSH into Azure VM}
\usage{
azr_ssh_cmd(
  name = make.names(Sys.info()["user"]),
  ...,
  ssh_dir = file.path(Sys.getenv("HOME"), ".ssh")
)
}
\arguments{
\item{name}{Resource group name}

\item{...}{Further rguments for \code{\link[=azr_vm]{azr_vm()}}.}

\item{ssh_dir}{The directory to store SSH files}
}
\description{
Command to SSH into Azure VM
}
