% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.R
\name{azr_ssh_priv}
\alias{azr_ssh_priv}
\title{Path to the Saved RSA Private Key}
\usage{
azr_ssh_priv(
  name = make.names(Sys.info()["user"]),
  ssh_dir = file.path(Sys.getenv("HOME"), ".ssh")
)
}
\arguments{
\item{name}{Resource group name}

\item{ssh_dir}{The directory to store SSH files}
}
\description{
Path to the Saved RSA Private Key
}
