% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_vm.R
\name{azr_vm}
\alias{azr_vm}
\title{Get Reference to VM}
\usage{
azr_vm(
  name = make.names(Sys.info()["user"]),
  dir = Sys.getenv("AZURE_DIR"),
  sub_id = Sys.getenv("AZURE_SUB_ID")
)
}
\arguments{
\item{name}{the name of the resource group. Defaults to your computer's username.}

\item{dir, sub_id}{Azure identifiers that should be set in your \code{.Renviron}. See
the package README for details.}
}
\description{
Get Reference to VM
}
