% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_vm.R
\name{azr_vm_setup}
\alias{azr_vm_setup}
\title{Set Up an Azure Virtual Machine for R}
\usage{
azr_vm_setup(
  name = make.names(Sys.info()["user"]),
  password = NULL,
  spatial = FALSE,
  rstudio = FALSE,
  size = "Standard_D2s_v5",
  loc = "eastus",
  dir = Sys.getenv("AZURE_DIR"),
  sub_id = Sys.getenv("AZURE_SUB_ID")
)
}
\arguments{
\item{name}{the name of the resource group. Will also be the username on the VM.}

\item{password}{the local VM password to set (will be the Rstudio password as well)}

\item{spatial}{if \code{TRUE} install spatial libraries as well}

\item{rstudio}{if \code{TRUE} install RStudio Server}

\item{size}{Machine specs. Some recommended options:
* \code{Standard_B2s}: 2 cores, 4 GB RAM
* \code{Standard_D2s_v5}: 2 cores, 8 GB RAM
* \code{Standard_D4s_v3}: 4 cores, 6 GB RAM
See \url{https://learn.microsoft.com/en-us/azure/virtual-machines/sizes} for more information.}

\item{loc}{The physical location of the VM. Defaults to \code{eastus} in Virginia.}

\item{dir, sub_id}{Azure identifiers that should be set in your \code{.Renviron}. See
the package README for details.}
}
\value{
A reference to the VM. See \code{AzureVM} package for details.
}
\description{
Set Up an Azure Virtual Machine for R
}
