% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{azure_authority_hosts}
\alias{azure_authority_hosts}
\title{Azure Authority Host URLs}
\usage{
azure_authority_hosts
}
\description{
Login endpoint URLs for different Azure cloud environments.
}
\keyword{internal}
