% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{azure_log_analytics_endpoints}
\alias{azure_log_analytics_endpoints}
\title{Azure Log Analytics Endpoint Defaults}
\format{
An object of class \code{list} of length 1.
}
\usage{
azure_log_analytics_endpoints
}
\description{
Default endpoint hosts for the Azure Log Analytics query API.
}
\keyword{internal}
