% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{azure_scopes}
\alias{azure_scopes}
\title{Common Azure OAuth Scopes}
\format{
An object of class \code{list} of length 4.
}
\usage{
azure_scopes
}
\description{
Predefined OAuth scopes for common Azure services.
}
\keyword{internal}
