% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{azure_storage_endpoints}
\alias{azure_storage_endpoints}
\title{Azure Storage Endpoint Defaults}
\format{
An object of class \code{list} of length 1.
}
\usage{
azure_storage_endpoints
}
\description{
Default endpoint suffixes for Azure Storage services.
}
\keyword{internal}
