% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credential.R
\name{Credential}
\alias{Credential}
\title{Azure Credential}
\description{
Base Credential class.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Credential-new}{\code{Credential$new()}}
\item \href{#method-Credential-clone}{\code{Credential$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Credential-new"></a>}}
\if{latex}{\out{\hypertarget{method-Credential-new}{}}}
\subsection{Method \code{new()}}{
Initialize the Credential
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Credential$new(
  scope = NULL,
  tenant_id = NULL,
  client_id = NULL,
  client_secret = NULL,
  cache_disk = TRUE,
  auth_host = "AZURE_PUBLIC_CLOUD",
  auth_url = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scope}}{The scope(s) to request access for}

\item{\code{client_id}}{The client ID of the application}

\item{\code{cache_disk}}{Whether to cache tokens to disk (default: TRUE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Credential} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Credential-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Credential-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Credential$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
