% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az-dataset.R
\name{dataset_uri}
\alias{dataset_uri}
\title{Build a URI for an \code{az_dataset}}
\usage{
dataset_uri(x, ...)
}
\arguments{
\item{x}{An \link{az_dataset} object.}

\item{tier}{Environment tier name (a key in \code{x@storage}).}

\item{uri_type}{URI type: \code{"https"} or \code{"hadoop"} (the Hadoop ABFS URI
form \verb{scheme://container@account.dfs.../path}, used by Spark, Flink,
Trino, and any \code{hadoop-azure} consumer).}
}
\value{
A character scalar URI.
}
\description{
Build a URI for an \code{az_dataset}
}
