% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_azure_client_secret}
\alias{default_azure_client_secret}
\title{Get default Azure client secret}
\usage{
default_azure_client_secret()
}
\value{
A character string with the client secret, or \code{NA_character_} if not set
}
\description{
Retrieves the Azure client secret from the \code{AZURE_CLIENT_SECRET} environment
variable, or returns \code{NA_character_} if not set.
}
\examples{
default_azure_client_secret()
}
