% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_azure_oauth_client}
\alias{default_azure_oauth_client}
\title{Create default Azure OAuth client}
\usage{
default_azure_oauth_client(
  client_id = default_azure_client_id(),
  client_secret = NULL,
  name = NULL
)
}
\arguments{
\item{client_id}{A character string specifying the client ID. Defaults to
\code{\link[=default_azure_client_id]{default_azure_client_id()}}.}

\item{client_secret}{A character string specifying the client secret. Defaults
to \code{NULL}.}

\item{name}{A character string specifying the client name. Defaults to \code{NULL}.}
}
\value{
An \code{\link[httr2:oauth_client]{httr2::oauth_client()}} object
}
\description{
Creates an \code{\link[httr2:oauth_client]{httr2::oauth_client()}} configured for Azure authentication.
}
\examples{
\dontrun{
client <- default_azure_oauth_client()
client <- default_azure_oauth_client(
  client_id = "my-client-id",
  client_secret = "my-secret"
)
}
}
