% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_federated_token_file}
\alias{default_federated_token_file}
\title{Get default federated token file path}
\usage{
default_federated_token_file()
}
\value{
A character string with the file path, or \code{NULL} if not set
}
\description{
Retrieves the path to the federated identity token file from the
\code{AZURE_FEDERATED_TOKEN_FILE} environment variable, or returns \code{NULL} if
not set. Used by \link{WorkloadIdentityCredential}.
}
\examples{
default_federated_token_file()
}
