% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_graph_endpoint}
\alias{default_graph_endpoint}
\title{Get default Microsoft Graph endpoint}
\usage{
default_graph_endpoint()
}
\value{
A character string with the Microsoft Graph endpoint host.
}
\description{
Returns the default host used to construct Microsoft Graph API URLs
(\code{graph.microsoft.com}).
}
\examples{
default_graph_endpoint()
}
