% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_log_analytics_endpoint}
\alias{default_log_analytics_endpoint}
\title{Get default Azure Log Analytics query endpoint}
\usage{
default_log_analytics_endpoint()
}
\value{
A character string with the Log Analytics query endpoint host.
}
\description{
Returns the default host used to construct Azure Log Analytics query URLs
(\code{api.loganalytics.io}).
}
\examples{
default_log_analytics_endpoint()
}
