% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_redirect_uri}
\alias{default_redirect_uri}
\title{Get default OAuth redirect URI}
\usage{
default_redirect_uri(redirect_uri = httr2::oauth_redirect_uri())
}
\arguments{
\item{redirect_uri}{A character string specifying the redirect URI. Defaults
to \code{\link[httr2:oauth_redirect_uri]{httr2::oauth_redirect_uri()}}.}
}
\value{
A character string with the redirect URI
}
\description{
Constructs a redirect URI for OAuth flows. If the provided URI doesn't have
a port, assigns a random port using \code{\link[httpuv:randomPort]{httpuv::randomPort()}}.
}
\examples{
default_redirect_uri()
}
