% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_refresh_token}
\alias{default_refresh_token}
\title{Get default Azure refresh token}
\usage{
default_refresh_token()
}
\value{
A character string with the refresh token, or \code{NULL} if not set
}
\description{
Retrieves the Azure refresh token from the \code{AZURE_REFRESH_TOKEN} environment
variable, or returns \code{NULL} if not set.
}
\examples{
default_refresh_token()
}
