% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_storage_endpoint}
\alias{default_storage_endpoint}
\title{Get default Azure Storage DFS endpoint suffix}
\usage{
default_storage_endpoint()
}
\value{
A character string with the DFS endpoint suffix.
}
\description{
Returns the default endpoint suffix used to construct Azure Data Lake Storage
Gen2 DFS URLs.
}
\examples{
default_storage_endpoint()
}
