% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default-credential.R
\name{print.credential_chain}
\alias{print.credential_chain}
\title{Print a credential chain}
\usage{
\method{print}{credential_chain}(x, ...)
}
\arguments{
\item{x}{A \code{credential_chain} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a credential chain
}
