% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azr-package.R
\docType{package}
\name{azr-package}
\alias{azr}
\alias{azr-package}
\title{azr: Configure Azure Services for R}
\description{
Provides functionality to install the Azure CLI, set up Azure virtual machines, and install the RStudio IDE on a VM instance.
}
\keyword{internal}
