% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azr-package.R
\docType{package}
\name{azr-package}
\alias{azr}
\alias{azr-package}
\title{azr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Pedro Baltazar \email{pedrobtz@gmail.com}

}
\keyword{internal}
