% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azr-package.R
\docType{package}
\name{azr-package}
\alias{azr}
\alias{azr-package}
\title{azr: Credential Chain for Seamless OAuth 2.0 Authentication to Azure Services}
\description{
Implements a credential chain for Azure OAuth 2.0 authentication based on 'httr2''s OAuth framework. Sequentially attempts authentication methods until one succeeds. During development allows interactive browser-based (Device Code and Auth Code Flows) login and non-interactive flows (Client Secret) in batch mode.
}
\author{
\strong{Maintainer}: Pedro Baltazar \email{pedrobtz@gmail.com}

}
\keyword{internal}
