% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{azr_options}
\alias{azr_options}
\title{azr package options}
\format{
An object of class \code{list} of length 2.
}
\usage{
azr_options
}
\description{
Registry of all azr package options. Each option can be set via an R option
or an environment variable. Use \code{\link[=azr_opt]{azr_opt()}} to read an option value.\tabular{lllll}{
   Name \tab R option \tab Env variable \tab Default \tab Description \cr
   \code{"verbose"} \tab \code{azr.verbose} \tab \code{AZR_VERBOSE} \tab \code{FALSE} \tab Enable verbose diagnostic output \cr
   \code{"cli_login_enable"} \tab \code{azr.cli_login_enable} \tab \code{AZR_CLI_LOGIN_ENABLE} \tab \code{FALSE} \tab Auto Azure CLI login \cr
}
}
\examples{
# List all option specs
azr_options

# Read with azr_opt()
azr_opt("verbose")

# Set for the session
options(azr.verbose = TRUE)

# Or via environment variable (before starting R)
# AZR_VERBOSE=true
}
\keyword{datasets}
