% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credential_client_secret.R
\name{ClientSecretCredential}
\alias{ClientSecretCredential}
\title{Client Secret Credential}
\description{
Authenticates using the authorization code flow to obtain access tokens.
}
\details{
This credential authenticates a user through the authorization code flow. The user must
navigate to a URL and enter a provided code to complete authentication.
Supports token caching to disk for persistence across sessions.
}
\section{Super class}{
\code{\link[azr:Credential]{azr::Credential}} -> \code{ClientSecretCredential}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ClientSecretCredential-validate}{\code{ClientSecretCredential$validate()}}
\item \href{#method-ClientSecretCredential-get_token}{\code{ClientSecretCredential$get_token()}}
\item \href{#method-ClientSecretCredential-req_auth}{\code{ClientSecretCredential$req_auth()}}
\item \href{#method-ClientSecretCredential-clone}{\code{ClientSecretCredential$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="azr" data-topic="Credential" data-id="initialize"><a href='../../azr/html/Credential.html#method-Credential-initialize'><code>azr::Credential$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="azr" data-topic="Credential" data-id="print"><a href='../../azr/html/Credential.html#method-Credential-print'><code>azr::Credential$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClientSecretCredential-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ClientSecretCredential-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClientSecretCredential$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClientSecretCredential-get_token"></a>}}
\if{latex}{\out{\hypertarget{method-ClientSecretCredential-get_token}{}}}
\subsection{Method \code{get_token()}}{
Get an access token using device code flow
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClientSecretCredential$get_token()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scope}}{Character vector of scopes to request (optional, uses initialized scope if not provided)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing the access token and expiration time
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClientSecretCredential-req_auth"></a>}}
\if{latex}{\out{\hypertarget{method-ClientSecretCredential-req_auth}{}}}
\subsection{Method \code{req_auth()}}{
Add authentication to an httr2 request
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClientSecretCredential$req_auth(req)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{req}}{An httr2 request object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An httr2 request object with bearer token authentication added
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClientSecretCredential-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ClientSecretCredential-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClientSecretCredential$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
