% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_azure_client_id}
\alias{default_azure_client_id}
\title{Get default Azure client ID}
\usage{
default_azure_client_id()
}
\value{
A character string with the client ID
}
\description{
Retrieves the Azure client ID from the \code{AZURE_CLIENT_ID} environment variable,
or falls back to the default Azure CLI client ID if not set.
}
\examples{
default_azure_client_id()
}
