% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_azure_tenant_id}
\alias{default_azure_tenant_id}
\title{Get default Azure tenant ID}
\usage{
default_azure_tenant_id()
}
\value{
A character string with the tenant ID
}
\description{
Retrieves the Azure tenant ID from the \code{AZURE_TENANT_ID} environment variable,
or falls back to the default value if not set.
}
\examples{
\dontrun{
default_azure_tenant_id()
}
}
