% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{environment_variables}
\alias{environment_variables}
\title{Azure Environment Variable Names}
\format{
An object of class \code{list} of length 9.
}
\usage{
environment_variables
}
\description{
Standard environment variable names used for Azure credential discovery.
}
\keyword{internal}
