% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credential-interactive.R
\name{InteractiveCredential}
\alias{InteractiveCredential}
\title{Interactive credential base class}
\description{
Base class for interactive authentication credentials. This class should not
be instantiated directly; use \link{DeviceCodeCredential} or \link{AuthCodeCredential}
instead.
}
\keyword{internal}
\section{Super class}{
\code{azr::Credential} -> \code{InteractiveCredential}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InteractiveCredential-is_interactive}{\code{InteractiveCredential$is_interactive()}}
\item \href{#method-InteractiveCredential-clone}{\code{InteractiveCredential$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="azr" data-topic="Credential" data-id="initialize"><a href='../../azr/html/Credential.html#method-Credential-initialize'><code>azr::Credential$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="azr" data-topic="Credential" data-id="print"><a href='../../azr/html/Credential.html#method-Credential-print'><code>azr::Credential$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="azr" data-topic="Credential" data-id="validate"><a href='../../azr/html/Credential.html#method-Credential-validate'><code>azr::Credential$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InteractiveCredential-is_interactive"></a>}}
\if{latex}{\out{\hypertarget{method-InteractiveCredential-is_interactive}{}}}
\subsection{Method \code{is_interactive()}}{
Check if the credential is interactive
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InteractiveCredential$is_interactive()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Always returns \code{TRUE} for interactive credentials
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InteractiveCredential-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InteractiveCredential-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InteractiveCredential$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
